work on callback tests.
